BETA: Fixing possible client crash due unbound buy of items (weight and inventory slots not validated). 	* Thanks Starter for report and first implementation. 	* This generic version implements exchangeItemsById(..) for information refers to its JavaDocs. 		* Try to use it instead of destroyItem/addItem methods. 	* Cleanup related to this changes.

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/itemcontainer/PcInventory.java
@@ -828,12 +828,12 @@ public class PcInventory extends Inventory
 	}
 	
 	@Override
-	public boolean validateCapacity(int slots)
+	public boolean validateCapacity(long slots)
 	{
 		return validateCapacity(slots, false);
 	}
 	
-	public boolean validateCapacity(int slots, boolean questItem)
+	public boolean validateCapacity(long slots, boolean questItem)
 	{
 		if (!questItem)
 			return (_items.size() - _questSlots + slots <= _owner.getInventoryLimit());
@@ -841,7 +841,7 @@ public class PcInventory extends Inventory
 	}
 	
 	@Override
-	public boolean validateWeight(int weight)
+	public boolean validateWeight(long weight)
 	{
 		// Disable weight check for GMs.
 		if (_owner.isGM() && _owner.getDietMode() && _owner.getAccessLevel().allowTransaction())
