+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/itemcontainer/PcWarehouse.java
@@ -47,7 +47,7 @@ public class PcWarehouse extends Warehouse
 	}
 	
 	@Override
-	public boolean validateCapacity(int slots)
+	public boolean validateCapacity(long slots)
 	{
 		return (_items.size() + slots <= _owner.getWareHouseLimit());
 	}
